Simplified codec changes.
